[zen-60] Refactored AnnotationManager, Scanner
- AnnotationManager now:
  - Composes an EventManager
  - createAnnotation() triggers the createAnnotation event, passing an
    event with "class", "content", and "raw" parameters representing the
    annotation requested. First listener to return an object halts the
    event, and that object is returned.
- AnnotationScanner adds a third element to the $annotation array
  representing the "raw" annotation. This array is then passed to
  createAnnotation(), and if a non-false value is returned, appended to
  the collection.
- Added GenericAnnotationParser, which operates in the style originally
  posited by Ralph.
